requirements.txt
google-cloud-aiplatform
google-cloud-firestore
requests
certifi
